requirements.txt
Flask==1.1.2
torch==1.6.0
torchvision==0.7.0
Pillow==7.2.0
beautifulsoup4==4.9.1
openimages==0.0.1
